requirements.txt
furl==2.1.3
sslyze==5.1.1
pyOpenSSL==23.1.0
pytest==7.2.1
pylint-exit==1.2.0
pytest-asyncio==0.20.3
ocsp-checker==1.9.11